refactor: 18049 Merge `VirtualRootNode` into `VirtualMap`
