+++ platform-sdk/platform-apps/tests/PlatformTestingTool/src/main/java/com/swirlds/demo/virtualmerkle/VirtualMerkleLeafHasher.java
@@ -170,7 +170,7 @@ public class VirtualMerkleLeafHasher {
             try {
                 // AccountVirtualMapKey -> AccountVirtualMapValue
                 final VirtualMap accountsMap = new VirtualMap(CONFIGURATION);
-                accountsMap.loadFromFile(roundFolder.resolve(accountsName), false);
+                accountsMap.loadFromFile(roundFolder.resolve(accountsName), false, false);
                 final VirtualMerkleLeafHasher accountsHasher = new VirtualMerkleLeafHasher(accountsMap);
                 accountsHash = accountsHasher.validate();
             } catch (final IOException e) {
@@ -180,7 +180,7 @@ public class VirtualMerkleLeafHasher {
             try {
                 // SmartContractMapKey -> SmartContractMapValue
                 final VirtualMap scMap = new VirtualMap(CONFIGURATION);
-                scMap.loadFromFile(roundFolder.resolve(scName), false);
+                scMap.loadFromFile(roundFolder.resolve(scName), false, false);
                 final VirtualMerkleLeafHasher scHasher = new VirtualMerkleLeafHasher(scMap);
                 scHash = scHasher.validate();
             } catch (final IOException e) {
@@ -190,7 +190,7 @@ public class VirtualMerkleLeafHasher {
             try {
                 // SmartContractByteCodeMapKey -> SmartContractByteCodeMapValue
                 final VirtualMap byteCodeMap = new VirtualMap(CONFIGURATION);
-                byteCodeMap.loadFromFile(roundFolder.resolve(scByteCodeName), false);
+                byteCodeMap.loadFromFile(roundFolder.resolve(scByteCodeName), false, false);
                 final VirtualMerkleLeafHasher byteCodeHasher = new VirtualMerkleLeafHasher(byteCodeMap);
                 byteCodeHash = byteCodeHasher.validate();
             } catch (final IOException e) {
