+++ platform-sdk/swirlds-virtualmap/src/test/java/com/swirlds/virtualmap/internal/merkle/VirtualInternalNodeTest.java
@@ -2,7 +2,7 @@
 package com.swirlds.virtualmap.internal.merkle;
 
 import static com.swirlds.virtualmap.internal.merkle.VirtualMapState.VM_STATE_KEY;
-import static com.swirlds.virtualmap.test.fixtures.VirtualMapTestUtils.createRoot;
+import static com.swirlds.virtualmap.test.fixtures.VirtualMapTestUtils.createMap;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -13,6 +13,7 @@ import static org.mockito.Mockito.mock;
 import com.swirlds.common.merkle.MerkleNode;
 import com.swirlds.common.merkle.route.MerkleRoute;
 import com.swirlds.common.test.fixtures.merkle.dummy.DummyBinaryMerkleInternal;
+import com.swirlds.virtualmap.VirtualMap;
 import com.swirlds.virtualmap.datasource.VirtualHashRecord;
 import com.swirlds.virtualmap.datasource.VirtualLeafBytes;
 import com.swirlds.virtualmap.test.fixtures.TestValueCodec;
@@ -35,8 +36,8 @@ class VirtualInternalNodeTest extends VirtualTestBase {
     @DisplayName("setChild is not a supported method")
     void setChildNotSupported() {
         final VirtualHashRecord virtualHashRecord = new VirtualHashRecord(1, null);
-        final VirtualRootNode root = createRoot();
-        final VirtualInternalNode internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        final VirtualMap map = createMap();
+        final VirtualInternalNode internalNode = new VirtualInternalNode(map, virtualHashRecord);
 
         final DummyBinaryMerkleInternal child = new DummyBinaryMerkleInternal();
         assertThrows(
@@ -58,8 +59,8 @@ class VirtualInternalNodeTest extends VirtualTestBase {
     @DisplayName("copy method is not supported")
     void copyNotSupported() {
         final VirtualHashRecord virtualHashRecord = new VirtualHashRecord(0, null);
-        final VirtualRootNode root = createRoot();
-        final VirtualInternalNode internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        final VirtualMap map = createMap();
+        final VirtualInternalNode internalNode = new VirtualInternalNode(map, virtualHashRecord);
 
         assertThrows(UnsupportedOperationException.class, internalNode::copy, "Copy is not supported");
     }
@@ -75,8 +76,8 @@ class VirtualInternalNodeTest extends VirtualTestBase {
         // (and in fact, any internal node other than the root node will ALWAYS have both a
         // left and right child).
         final VirtualHashRecord virtualHashRecord = new VirtualHashRecord(1, null);
-        final VirtualRootNode root = createRoot();
-        final VirtualInternalNode internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        final VirtualMap map = createMap();
+        final VirtualInternalNode internalNode = new VirtualInternalNode(map, virtualHashRecord);
         assertNull(internalNode.getChild(0), "No left child");
         assertNull(internalNode.getChild(1), "No right child");
     }
@@ -92,15 +93,15 @@ class VirtualInternalNodeTest extends VirtualTestBase {
         // that I know that under normal circumstances the call would succeed, and is only *not*
         // returning a valid because it shouldn't.
         final VirtualHashRecord virtualHashRecord = new VirtualHashRecord(3, null);
-        final VirtualRootNode root = createRoot();
-        root.put(A_KEY, APPLE, TestValueCodec.INSTANCE);
-        root.put(B_KEY, BANANA, TestValueCodec.INSTANCE);
-        root.put(C_KEY, CHERRY, TestValueCodec.INSTANCE);
-        root.put(D_KEY, DATE, TestValueCodec.INSTANCE);
-        root.put(E_KEY, EGGPLANT, TestValueCodec.INSTANCE);
-        root.put(F_KEY, FIG, TestValueCodec.INSTANCE);
-        root.put(G_KEY, GRAPE, TestValueCodec.INSTANCE);
-        final VirtualInternalNode internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        final VirtualMap map = createMap();
+        map.put(A_KEY, APPLE, TestValueCodec.INSTANCE);
+        map.put(B_KEY, BANANA, TestValueCodec.INSTANCE);
+        map.put(C_KEY, CHERRY, TestValueCodec.INSTANCE);
+        map.put(D_KEY, DATE, TestValueCodec.INSTANCE);
+        map.put(E_KEY, EGGPLANT, TestValueCodec.INSTANCE);
+        map.put(F_KEY, FIG, TestValueCodec.INSTANCE);
+        map.put(G_KEY, GRAPE, TestValueCodec.INSTANCE);
+        final VirtualInternalNode internalNode = new VirtualInternalNode(map, virtualHashRecord);
         final VirtualLeafNode leftChild = internalNode.getChild(0);
         final VirtualLeafNode rightChild = internalNode.getChild(1);
         assertNotNull(leftChild, "child should not be null");
@@ -114,21 +115,21 @@ class VirtualInternalNodeTest extends VirtualTestBase {
     @Tags({@Tag("VirtualMerkle")})
     @DisplayName("Getting a child that isn't 0 or 1")
     void getInternalNodeFromCacheAndFromDisk() {
-        final VirtualRootNode root = createRoot();
-        root.put(A_KEY, APPLE, TestValueCodec.INSTANCE);
-        root.put(B_KEY, BANANA, TestValueCodec.INSTANCE);
-        root.put(C_KEY, CHERRY, TestValueCodec.INSTANCE);
-        root.put(D_KEY, DATE, TestValueCodec.INSTANCE);
-        root.put(E_KEY, EGGPLANT, TestValueCodec.INSTANCE);
-        root.put(F_KEY, FIG, TestValueCodec.INSTANCE);
-        root.put(G_KEY, GRAPE, TestValueCodec.INSTANCE);
+        final VirtualMap map = createMap();
+        map.put(A_KEY, APPLE, TestValueCodec.INSTANCE);
+        map.put(B_KEY, BANANA, TestValueCodec.INSTANCE);
+        map.put(C_KEY, CHERRY, TestValueCodec.INSTANCE);
+        map.put(D_KEY, DATE, TestValueCodec.INSTANCE);
+        map.put(E_KEY, EGGPLANT, TestValueCodec.INSTANCE);
+        map.put(F_KEY, FIG, TestValueCodec.INSTANCE);
+        map.put(G_KEY, GRAPE, TestValueCodec.INSTANCE);
 
-        MerkleNode child = root.getChild(1);
+        MerkleNode child = map.getChild(1);
         assertTrue(child instanceof VirtualInternalNode, "child should be an internal node");
         VirtualInternalNode internalNode = (VirtualInternalNode) child;
 
         assertEquals(5, ((VirtualInternalNode) internalNode.getChild(0)).getPath());
-        root.getCache().release();
+        map.getCache().release();
         assertEquals(5, ((VirtualInternalNode) internalNode.getChild(0)).getPath());
     }
 
@@ -140,9 +141,9 @@ class VirtualInternalNodeTest extends VirtualTestBase {
     @Tags({@Tag("VirtualMerkle")})
     @DisplayName("Getting a child that is on disk")
     void getValidChildOnDisk() throws IOException {
-        final VirtualRootNode root = createRoot();
-        root.getState().setLastLeafPath(14);
-        root.getState().setFirstLeafPath(7);
+        final VirtualMap map = createMap();
+        map.getState().setLastLeafPath(14);
+        map.getState().setFirstLeafPath(7);
 
         final List<VirtualLeafBytes> leaves = List.of(
                 new VirtualLeafBytes<>(8, D_KEY, DATE, TestValueCodec.INSTANCE),
@@ -152,16 +153,16 @@ class VirtualInternalNodeTest extends VirtualTestBase {
                 new VirtualLeafBytes<>(12, F_KEY, FIG, TestValueCodec.INSTANCE),
                 new VirtualLeafBytes<>(13, C_KEY, CHERRY, TestValueCodec.INSTANCE),
                 new VirtualLeafBytes<>(14, G_KEY, GRAPE, TestValueCodec.INSTANCE));
-        root.getDataSource().saveRecords(7, 14, leaves.stream().map(this::hash), leaves.stream(), Stream.empty());
+        map.getDataSource().saveRecords(7, 14, leaves.stream().map(this::hash), leaves.stream(), Stream.empty());
 
         VirtualHashRecord virtualHashRecord = new VirtualHashRecord(3, null);
-        VirtualInternalNode internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        VirtualInternalNode internalNode = new VirtualInternalNode(map, virtualHashRecord);
         VirtualLeafNode rightChildLeaf = internalNode.getChild(1);
         assertNotNull(rightChildLeaf, "value should not be null");
         assertEquals(D_KEY, rightChildLeaf.getKey(), "key should match original");
 
         virtualHashRecord = new VirtualHashRecord(4, null);
-        internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        internalNode = new VirtualInternalNode(map, virtualHashRecord);
         VirtualLeafNode leftChildLeaf = internalNode.getChild(0);
         rightChildLeaf = internalNode.getChild(1);
         assertNotNull(leftChildLeaf, "value should not be null");
@@ -170,7 +171,7 @@ class VirtualInternalNodeTest extends VirtualTestBase {
         assertEquals(E_KEY, rightChildLeaf.getKey(), "key should match original");
 
         virtualHashRecord = new VirtualHashRecord(5, null);
-        internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        internalNode = new VirtualInternalNode(map, virtualHashRecord);
         leftChildLeaf = internalNode.getChild(0);
         rightChildLeaf = internalNode.getChild(1);
         assertNotNull(leftChildLeaf, "value should not be null");
@@ -179,7 +180,7 @@ class VirtualInternalNodeTest extends VirtualTestBase {
         assertEquals(F_KEY, rightChildLeaf.getKey(), "key should match original");
 
         virtualHashRecord = new VirtualHashRecord(6, null);
-        internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        internalNode = new VirtualInternalNode(map, virtualHashRecord);
         leftChildLeaf = internalNode.getChild(0);
         rightChildLeaf = internalNode.getChild(1);
         assertNotNull(leftChildLeaf, "value should not be null");
@@ -193,8 +194,8 @@ class VirtualInternalNodeTest extends VirtualTestBase {
     @DisplayName("Check version and class are valid")
     void getVersionAndClassId() {
         final VirtualHashRecord virtualHashRecord = new VirtualHashRecord(0, null);
-        final VirtualRootNode root = createRoot();
-        final VirtualInternalNode internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        final VirtualMap map = createMap();
+        final VirtualInternalNode internalNode = new VirtualInternalNode(map, virtualHashRecord);
         assertEquals(0xaf2482557cfdb6bfL, internalNode.getClassId(), "Increases code coverage for getClassId");
         assertEquals(1, internalNode.getVersion(), "Increases code coverage for getVersion");
     }
@@ -205,8 +206,8 @@ class VirtualInternalNodeTest extends VirtualTestBase {
     void toStringTest() {
         // Honestly, I'm just juicing the code coverage
         final VirtualHashRecord virtualHashRecord = new VirtualHashRecord(0, null);
-        final VirtualRootNode root = createRoot();
-        final VirtualInternalNode internalNode = new VirtualInternalNode(root, virtualHashRecord);
+        final VirtualMap map = createMap();
+        final VirtualInternalNode internalNode = new VirtualInternalNode(map, virtualHashRecord);
         assertNotNull(internalNode.toString(), "value should not be null");
     }
 }
