+++ platform-sdk/swirlds-virtualmap/src/testFixtures/java/com/swirlds/virtualmap/test/fixtures/VirtualMapTestUtils.java
@@ -8,8 +8,6 @@ import com.swirlds.config.api.ConfigurationBuilder;
 import com.swirlds.virtualmap.VirtualMap;
 import com.swirlds.virtualmap.config.VirtualMapConfig;
 import com.swirlds.virtualmap.datasource.VirtualDataSourceBuilder;
-import com.swirlds.virtualmap.internal.merkle.VirtualMapState;
-import com.swirlds.virtualmap.internal.merkle.VirtualRootNode;
 
 /**
  * Methods for testing {@link VirtualMap}.
@@ -35,20 +33,5 @@ public final class VirtualMapTestUtils {
 
     public static VirtualMap createMap() {
         return createMap(VM_LABEL);
-    }
-
-    public static VirtualRootNode createRoot() {
-        return createRoot(CONFIGURATION);
-    }
-
-    public static VirtualRootNode createRoot(final Configuration configuration) {
-        final VirtualRootNode root =
-                new VirtualRootNode(new InMemoryBuilder(), configuration.getConfigData(VirtualMapConfig.class));
-        root.postInit(new VirtualMapState(VM_LABEL));
-        return root;
-    }
-
-    public static VirtualRootNode getRoot(VirtualMap map) {
-        return map.getChild(0);
     }
 }
